wip useEffect commented out in applist page
